live like dislike issue fixed and livestream api added details

+++ database/migrations/2022_08_08_062315_add_like_dislikes_to_like_dislikes_table.php
@@ -14,7 +14,7 @@ class AddLikeDislikesToLikeDislikesTable extends Migration
     public function up()
     {
         Schema::table('like_dislikes', function (Blueprint $table) {
-            $table->tinyInteger('live_id')->nullable()->after('audio_id');
+            $table->integer('live_id')->nullable()->after('audio_id');
 
         });
     }
